Deployment/requirements.txt
streamlit
tensorflow
keras
numpy
pillow 
urllib3